+++ packages/core/src/index.ts
@@ -3,5 +3,4 @@ export * from "./component";
 export * from "./function";
 export * from "./hook";
 export * from "./jsx";
-export * from "./ref";
 export * from "./semantic";