+++ packages/plugins/eslint-plugin-react-x/src/rules/refs/refs.ts
@@ -1,8 +1,11 @@
 import * as ast from "@eslint-react/ast";
 import * as core from "@eslint-react/core";
+import { isUseRefCall } from "@eslint-react/core";
 import { type RuleContext, type RuleFeature, defineRuleListener } from "@eslint-react/shared";
+import type { Scope } from "@typescript-eslint/scope-manager";
 import { AST_NODE_TYPES as AST } from "@typescript-eslint/types";
 import type { TSESTree } from "@typescript-eslint/utils";
+import { findVariable } from "@typescript-eslint/utils/ast-utils";
 import { P, isMatching, match } from "ts-pattern";
 
 import { createRule } from "../../utils";
@@ -83,6 +86,26 @@ export function create(context: RuleContext<MessageID, []>) {
     return checkSides(left, right) || checkSides(right, left);
   }
 
+  function isInitializedFromRef(name: string, initialScope: Scope) {
+    for (const { node } of findVariable(initialScope, name)?.defs ?? []) {
+      if (node.type !== AST.VariableDeclarator) continue;
+      const init = node.init;
+      if (init == null) continue;
+      switch (true) {
+        // const identifier = anotherRef.current;
+        case init.type === AST.MemberExpression
+          && init.object.type === AST.Identifier
+          && (init.object.name === "ref" || init.object.name.endsWith("Ref")):
+          return true;
+        // const identifier = useRef();
+        case init.type === AST.CallExpression
+          && isUseRefCall(init):
+          return true;
+      }
+    }
+    return false;
+  }
+
   return defineRuleListener(
     hCollector.visitor,
     cCollector.visitor,
@@ -138,9 +161,9 @@ export function create(context: RuleContext<MessageID, []>) {
           const obj = node.object;
           if (obj.type !== AST.Identifier) continue;
           switch (true) {
-            case core.isRefLikeName(obj.name):
+            case obj.name === "ref" || obj.name.endsWith("Ref"):
             case jsxRefIdentifiers.has(obj.name):
-            case core.isInitializedFromRef(obj.name, context.sourceCode.getScope(node.object)):
+            case isInitializedFromRef(obj.name, context.sourceCode.getScope(node.object)):
               break;
             default:
               continue;